Use the editor's .import directory instead of a cache folder
The Editor class EditorSceneImporter has a function
import_scene_from_other_importer which is documented in the code as a
utility for importers (like fbx2gltf) to call other scene importers
(like Godot's glTF 2.0 importer). It loads the scene and provides a Node
that can be saved to a PackedScene. This avoids the need to manually
load the glTF file using ResourceLoader, and lets us use the import
cache directly